+++ duckbot/slash/context.py
@@ -1,0 +1,51 @@
+from discord.ext.commands.view import StringView
+from .route import Route8
+
+
+class InteractionContext:
+    def __init__(self, *, bot, interaction, command):
+        self.bot = bot
+        self.interaction = interaction
+        self.channel = interaction.channel
+        self.guild = interaction.guild
+        self.author = interaction.author
+        self.command = command
+        self.invoked_with = None
+        self.invoked_parents = []
+        self.invoked_subcommand = None
+        self.subcommand_passed = None
+        self.command_failed = False
+        self._state = bot._connection
+        options = interaction.data.get("options", [])
+        if options and options[0] and not options[0].get("value", None):
+            if command._slash_discordpy_adapt_name[options[0]["name"]]:
+                self.view = options[0]["name"] + " " + " ".join([x.get("value", "") for x in options[0].get("options", [])])
+            else:
+                self.view = " ".join([x.get("value", "") for x in options[0].get("options", [])])
+        else:
+            self.view = " ".join([x.get("value", "") for x in options])
+        self.view = StringView(self.view)
+        self.follow_up = False
+
+    async def send(self, content="", *, embed=None):
+        if self.follow_up:
+            json = {"content": content, "embeds": [embed.to_dict()] if embed else [], "tts": False}
+            route = Route8("POST", f"/webhooks/{self.interaction.application_id}/{self.interaction.token}")
+        else:
+            json = {"type": 4, "data": {"content": content, "embeds": [embed.to_dict()] if embed else [], "tts": False}}
+            route = Route8("POST", f"/interactions/{self.interaction.id}/{self.interaction.token}/callback")
+        await self.bot.http.request(route, json=json)
+
+    def typing(self):
+        return self
+
+    async def __aenter__(self):
+        assert not self.follow_up
+        json = {"type": 5, "data": {"content": ":thinking:"}}
+        route = Route8("POST", f"/interactions/{self.interaction.id}/{self.interaction.token}/callback")
+        await self.bot.http.request(route, json=json)
+        self.follow_up = True
+        return self
+
+    async def __aexit__(self, exc_type, exc, tb):
+        return None
