+++ duckbot/slash/route.py
@@ -1,0 +1,5 @@
+from discord.http import Route
+
+
+class Route8(Route):
+    BASE = "https://discord.com/api/v8"
